Resolved conflict in database/seeders/PaymentLibrariesSeeder.php
keeping new gateways 66 (LawPay) and 67 (CHIP) visible.

+++ app/PaymentDrivers/StripePaymentDriver.php
@@ -26,7 +26,7 @@ use Stripe\PaymentIntent;
 use Stripe\PaymentMethod;
 use App\Models\GatewayType;
 use App\Models\PaymentHash;
-use App\Http\Requests\Request;
+use Illuminate\Http\Request;
 use App\Jobs\Util\SystemLogger;
 use App\Utils\Traits\MakesHash;
 use App\Exceptions\PaymentFailed;
@@ -272,12 +272,12 @@ class StripePaymentDriver extends BaseDriver implements SupportsHeadlessInterfac
             $this->client
            && isset($this->client->country)
            && (
-                (in_array($this->client->country->iso_3166_2, ['FR', 'IE', 'NL', 'DE', 'ES']) && $this->client->currency()->code == 'EUR')
+               (in_array($this->client->country->iso_3166_2, ['FR', 'IE', 'NL', 'DE', 'ES']) && $this->client->currency()->code == 'EUR')
                 || ($this->client->country->iso_3166_2 == 'JP' && $this->client->currency()->code == 'JPY')
                 || ($this->client->country->iso_3166_2 == 'MX' && $this->client->currency()->code == 'MXN')
                 || ($this->client->country->iso_3166_2 == 'GB' && $this->client->currency()->code == 'GBP')
                 || ($this->client->country->iso_3166_2 == 'US' && $this->client->currency()->code == 'USD')
-            )
+           )
         ) {
             $types[] = GatewayType::DIRECT_DEBIT;
         }
@@ -766,25 +766,7 @@ class StripePaymentDriver extends BaseDriver implements SupportsHeadlessInterfac
         if ($request->type === 'charge.succeeded') {
             foreach ($request->data as $transaction) {
 
-                $payment = Payment::query()
-                    ->where('company_id', $this->company_gateway->company_id)
-                    ->where(function ($query) use ($transaction) {
-
-                        if (isset($transaction['payment_intent'])) {
-                            $query->where('transaction_reference', $transaction['payment_intent']);
-                        }
-
-                        if (isset($transaction['payment_intent']) && isset($transaction['id'])) {
-                            $query->orWhere('transaction_reference', $transaction['id']);
-                        }
-
-                        if (!isset($transaction['payment_intent']) && isset($transaction['id'])) {
-                            $query->where('transaction_reference', $transaction['id']);
-                        }
-
-                    })
-                    ->first();
-
+                $payment = self::findPaymentByStripeReference($this->company_gateway->company_id, $transaction);
 
                 if ($payment) {
 
@@ -812,27 +794,7 @@ class StripePaymentDriver extends BaseDriver implements SupportsHeadlessInterfac
 
                 if ($charge->captured) {
 
-
-                    $payment = Payment::query()
-                        ->where('company_id', $this->company_gateway->company_id)
-                        ->where(function ($query) use ($transaction) {
-
-                            if (isset($transaction['payment_intent'])) {
-                                $query->where('transaction_reference', $transaction['payment_intent']);
-                            }
-
-                            if (isset($transaction['payment_intent']) && isset($transaction['id'])) {
-                                $query->orWhere('transaction_reference', $transaction['id']);
-                            }
-
-                            if (!isset($transaction['payment_intent']) && isset($transaction['id'])) {
-                                $query->where('transaction_reference', $transaction['id']);
-                            }
-
-                        })
-                        ->first();
-
-
+                    $payment = self::findPaymentByStripeReference($this->company_gateway->company_id, $transaction);
 
                     if ($payment) {
                         $payment->status_id = Payment::STATUS_COMPLETED;
@@ -861,12 +823,11 @@ class StripePaymentDriver extends BaseDriver implements SupportsHeadlessInterfac
 
                 return response()->json([], 200);
             } elseif ($request->data['object']['status'] == "inactive" && $request->data['object']['payment_method']) {
-                // Delete payment method
                 $clientgateway = ClientGatewayToken::query()
                     ->where('token', $request->data['object']['payment_method'])
                     ->first();
 
-                if ($clientgateway) {
+                if ($clientgateway && !str_starts_with($clientgateway->token, 'ba_')) { //ba_ tokens should not be deleted
                     $clientgateway->delete();
                 }
 
@@ -921,7 +882,6 @@ class StripePaymentDriver extends BaseDriver implements SupportsHeadlessInterfac
      * https://stripe.com/docs/api/payment_methods/detach
      *
      * @param ClientGatewayToken $token
-     * @return void
      */
     public function detach(ClientGatewayToken $token)
     {
@@ -945,6 +905,8 @@ class StripePaymentDriver extends BaseDriver implements SupportsHeadlessInterfac
                 $this->client->company
             );
         }
+
+        return true;
     }
 
     public function getCompanyGatewayId(): int
